updated CRV API
