+++ src/main/scala/chiselverify/crv/Rand.scala
@@ -19,37 +19,37 @@ trait Rand {
     * @param that a second parameter for the addition constraint
     * @return rand variable being the result of the addition constraint.
     */
-  def #+(that: U): U
+  def +(that: U): U
 
   /** defines the add constraint between a [[Rand]] variable and a BigInt.
     * @param that a second parameter for the addition constraint
     * @return rand variable being the result of the addition constraint.
     */
-  def #+(that: BigInt): U
+  def +(that: BigInt): U
 
   /** defines the subtraction constraint between two [[Rand]] variables.
     * @param that a second parameter for the addition constraint
     * @return rand variable being the result of the addition constraint.
     */
-  def #-(that: U): U
+  def -(that: U): U
 
   /** defines the subtraction constraint between a [[Rand]] variable and a BigInt.
     * @param that a second parameter for the addition constraint
     * @return rand variable being the result of the addition constraint.
     */
-  def #-(that: BigInt): U
+  def -(that: BigInt): U
 
   /** defines the multiplication constraint between two [[Rand]] variables.
     * @param that a second parameter for the addition constraint
     * @return rand variable being the result of the addition constraint.
     */
-  def #*(that: U): U
+  def *(that: U): U
 
   /** defines the multiplication constraint between a [[Rand]] variable and a BigInt.
     * @param that a second parameter for the addition constraint
     * @return rand variable being the result of the addition constraint.
     */
-  def #*(that: BigInt): U
+  def *(that: BigInt): U
 
   /** defines the division constraint between two [[Rand]] variables.
     * @param that a second parameter for the addition constraint
@@ -79,83 +79,83 @@ trait Rand {
     * @param that a second parameter for the addition constraint
     * @return rand variable being the result of the addition constraint.
     */
-  def #^(that: U): U
+  def ^(that: U): U
 
   /** defines the exponential constraint between a [[Rand]] variable and a BigInt.
     * @param that a second parameter for the addition constraint
     * @return rand variable being the result of the addition constraint.
     */
-  def #^(that: BigInt): U
+  def ^(that: BigInt): U
 
   /** Defines inequality [[Constraint]] between [[Rand]] and BigInt constant.
     *
     * @param that a second parameter for inequality [[Constraint]].
     * @return the defined [[Constraint]].
     */
-  def #\=(that: U): Constraint
+  def \=(that: U): Constraint
 
   /** Defines inequality [[Constraint]] between [[Rand]] and BigInt constant.
     *
     * @param that a second parameter for inequality [[Constraint]].
     * @return the defined [[Constraint]].
     */
-  def #\=(that: BigInt): Constraint
+  def \=(that: BigInt): Constraint
 
   /** Defines "less than" [[Constraint]] between two [[Rand]].
     *
     * @param that a second parameter for "less than" [[Constraint]].
     * @return the defined [[Constraint]].
     */
-  def #<(that: U): Constraint
+  def <(that: U): Constraint
 
   /** Defines "less than" [[Constraint]] between [[Rand]] and BigInt constant.
     *
     * @param that a second parameter for "less than" [[Constraint]].
     * @return the equation [[Constraint]].
     */
-  def #<(that: BigInt): Constraint
+  def <(that: BigInt): Constraint
 
   /** Defines "less than or equal" [[Constraint]] between two [[Rand]].
     *
     * @param that a second parameter for "less than or equal" [[Constraint]].
     * @return the defined [[Constraint]].
     */
-  def #<=(that: U): Constraint
+  def <=(that: U): Constraint
 
   /** Defines "less than or equal" [[Constraint]] between [[Rand]] and BigInt constant.
     *
     * @param that a second parameter for "less than or equal" [[Constraint]].
     * @return the equation [[Constraint]].
     */
-  def #<=(that: BigInt): Constraint
+  def <=(that: BigInt): Constraint
 
   /** Defines "greater than" [[Constraint]] between two [[Rand]].
     *
     * @param that a second parameter for "greater than or equal" [[Constraint]].
     * @return the defined [[Constraint]].
     */
-  def #>(that: U): Constraint
+  def >(that: U): Constraint
 
   /** Defines "greater than" [[Constraint]] between [[Rand]] and BigInt constant.
     *
     * @param that a second parameter for "greater than or equal" [[Constraint]].
     * @return the equation [[Constraint]].
     */
-  def #>(that: BigInt): Constraint
+  def >(that: BigInt): Constraint
 
   /** Defines "greater than or equal" [[Constraint]] between two [[Rand]].
     *
     * @param that a second parameter for "greater than or equal" [[Constraint]].
     * @return the defined [[Constraint]].
     */
-  def #>=(that: U): Constraint
+  def >=(that: U): Constraint
 
   /** Defines "greater than or equal" [[Constraint]] between [[Rand]] and BigInt constant.
     *
     * @param that a second parameter for "greater than or equal" [[Constraint]].
     * @return the equation [[Constraint]].
     */
-  def #>=(that: BigInt): Constraint
+  def >=(that: BigInt): Constraint
 
   /** Set the value of the current random variable
     * @param that the value to assign to the random variable
