+++ src/test/scala/examples/leros/AluAccuTester.scala
@@ -20,13 +20,7 @@ class AluAccuTester extends FlatSpec with ChiselScalatestTester with Matchers {
                 bin("lo10", 0 to 10),
                 bin("First100", 0 to 100)
             )
-                // CoverPoint(dut.io.test, "test", //CoverPoint 2
-                //     Bins("testLo10", 0 to 10)::Nil)::
         )
-        //Declare cross points
-        /*Cross("accuAndTest", "accu", "test",
-            CrossBin("both1", 1 to 1, 1 to 1)::Nil)::
-        Nil)*/
 
         def alu(a: Int, b: Int, op: Int): Int = {
